combine regression e2e with verifier test
